Make vim & tmux colors dynamic
Based on the Terminal profile that is used. So I can either use a dark
or a light BG version (I like switching depending on the light around
me), I just need to make sure that TBG & TFB env vars are set.

Whoop whoop! I wanted to have this for a while, only finally sat down to
set this up. I should probably write a blog post or something.